scripts/requirements.txt
nomic
supabase
python-dotenv
tqdm
requests
